+++ packages/f-header/test-utils/component-objects/f-header.selectors.js
@@ -1,0 +1,8 @@
+export const HEADER_COMPONENT = '[data-test-id="header-component"]'
+
+const loginLink = () => $('[data-test-id="login-link"]');
+const offersLink = () => $$('[data-test-id="offers-link"]');
+const deliveryEnquiryLink = () => $('[data-test-id="delivery-link"]');
+const helpLink = () => $('[data-test-id="help-link"]');
+const headerLogo = () => $('[data-test-id="header-logo"]');
+const mobileNavigation = () => $('[data-test-id="nav-toggle"]');
